refactor test archives
