+++ blockchain/biobank-contract/package.json
@@ -1,6 +1,6 @@
 {
     "name": "biobank",
-    "version": "11.0.0",
+    "version": "1.0.0",
     "description": "Biobank contract implemented in JavaScript",
     "main": "index.js",
     "engines": {
