+++ blockchain/commands.sh
@@ -48,11 +48,14 @@ export CORE_PEER_MSPCONFIGPATH=${PWD}/organizations/peerOrganizations/org2.examp
 export CORE_PEER_ADDRESS=localhost:9051
 
 
-# -------------------------------- IBM BLOCKCHAIN
+# -------------------------------- IBM BLOCKCHAIN Microfab--------------------------
 
+# Start Microfab
 START_IMAGE="ibmcom/ibp-microfab:0.0.12"
 docker run -e MICROFAB_CONFIG --label fabric-environment-name="1 Org Local Fabric Microfab" -p 8080:8080 $START_IMAGE
 
+
+# clear Microfab
 CONTAINER=$(docker ps -f label=fabric-environment-name="1 Org Local Fabric Microfab" -q -a)
 for CONTAINER in $(docker ps -f label=fabric-environment-name="1 Org Local Fabric Microfab" -q -a) 
 do 
@@ -61,14 +64,11 @@ done
 
 docker volume prune -f
 
-# for VOLUME in $(docker volume ls -f label=fabric-environment-name="1 Org Local Fabric Microfab" -q)
-# do 
-# docker volume rm -f ${VOLUME} 
-# done
-
-
+# Microfab config - 1 channel
 export MICROFAB_CONFIG='{"port":8080,  "endorsing_organizations": [{"name": "Org1"}],"channels": [{"name": "mychannel","endorsing_organizations": ["Org1"]}]}'
 
+
+# Microfab config - 2 channel
 export MICROFAB_CONFIG='{
     "port": 8080,
     "endorsing_organizations":[
@@ -95,21 +95,3 @@ export MICROFAB_CONFIG='{
     "timeout": "60s"
 }'
 
-export MICROFAB_CONFIG='{
-    "port": 8080,
-    "endorsing_organizations":[
-        {
-            "name": "Org1"
-        }
-      ],
-    "channels":[
-        {
-            "name": "channel1",
-            "endorsing_organizations":[
-                "Org1"
-            ],
-            "capability_level": "V2_0"
-        }
-    ],
-    "timeout": "60s"
-}'
