+++ blockchain/currency-contract/functionalTests/BiocoinContract-biobank.test.js
@@ -1,23 +1,3 @@
-/*
-* Use this file for functional testing of your smart contract.
-* Fill out the arguments and return values for a function and
-* use the CodeLens links above the transaction blocks to
-* invoke/submit transactions.
-* All transactions defined in your smart contract are used here
-* to generate tests, including those functions that would
-* normally only be used on instantiate and upgrade operations.
-* This basic test file can also be used as the basis for building
-* further functional tests to run as part of a continuous
-* integration pipeline, or for debugging locally deployed smart
-* contracts by invoking/submitting individual transactions.
-*/
-/*
-* Generating this test file will also trigger an npm install
-* in the smart contract project directory. This installs any
-* package dependencies, including fabric-network, which are
-* required for this test file to be run locally.
-*/
-
 'use strict';
 
 const assert = require('assert');
@@ -26,17 +6,16 @@ const SmartContractUtil = require('./js-smart-contract-util');
 const TestAccountUtil = require('./test-utils/test-account-util')
 const os = require('os');
 const path = require('path');
+const CONFIG = require('./config.json');
 const { Test } = require('mocha');
 const { doesNotMatch } = require('assert');
 
-describe('BiocoinContract-biobank@8.0.0' , () => {
-
+describe('BiocoinContract-biobank' , () => {
     const homedir = os.homedir();
-    const walletPath = path.join(homedir, '.fabric-vscode', 'environments', 'testNetwork', 'wallets', 'Org1');
+    const walletPath = path.join(homedir, CONFIG.walletPath);
+    const identityName = CONFIG.identityName
     const gateway = new fabricNetwork.Gateway();
     let wallet;
-    // const identityName = 'Org1 Admin';
-    const identityName = 'user';
     let connectionProfile;
 
     before(async () => {
@@ -45,7 +24,6 @@ describe('BiocoinContract-biobank@8.0.0' , () => {
     });
 
     beforeEach(async () => {
-
         const discoveryAsLocalhost = SmartContractUtil.hasLocalhostURLs(connectionProfile);
         const discoveryEnabled = true;
 
@@ -75,7 +53,7 @@ describe('BiocoinContract-biobank@8.0.0' , () => {
             const arg1 = TestAccountUtil.anotherGeneratedAddress;
             const arg2 = 10;
             const args = [ arg0, arg1, arg2];
-            const response = await SmartContractUtil.submitTransaction('BiocoinContract', 'transferBiocoins', args, gateway); // Returns buffer of transaction return value
+            const response = await SmartContractUtil.submitTransaction('BiocoinContract', 'transferBiocoins', args, gateway);
             
             const json_response = JSON.parse(response.toString())
             assert.strictEqual(json_response[0]['balance'], 0);
@@ -91,7 +69,7 @@ describe('BiocoinContract-biobank@8.0.0' , () => {
             const arg1 = TestAccountUtil.anotherGeneratedAddress;
             const arg2 = 5.5;
             const args = [ arg0, arg1, arg2];
-            const response = await SmartContractUtil.submitTransaction('BiocoinContract', 'transferBiocoins', args, gateway); // Returns buffer of transaction return value
+            const response = await SmartContractUtil.submitTransaction('BiocoinContract', 'transferBiocoins', args, gateway);
             
             const json_response = JSON.parse(response.toString())
             assert.strictEqual(json_response[0]['balance'], 4.5);
