+++ blockchain/currency-contract/functionalTests/config.json
@@ -1,0 +1,9 @@
+{
+  "walletPath":  ".fabric-vscode/environments/testNetwork/wallets/Org1",
+  "identityName": "Org1 Admin",
+  "connectionProfilePath":".fabric-vscode/v2/environments/1 Org Local Fabric/gateways/Org1 Gateway.json",
+
+  "__walletPath": "Desktop/toshi/poli/biobank/application/fabric-details/wallet",
+  "__identityName": "user",
+  "__connectionProfilePath": "Desktop/poli/biobank/blockchain/test-network/organizations/peerOrganizations/org1.example.com/connection-org1.json"
+}
