+++ blockchain/success_commands.sh
